Fixed tests.
Fixed some of the tests. Disabled crAddVideo_crash as the parent process
remains waiting for the completion event for some reason. This is a
separate issue with the event not firing.

Disabled delivery tests as they depend on local configuration. For these
to be enabled again, some CMake Variables need to be introduced where
the local configuration can be set - including which tests to check.
Also they do not really make sense if the arrival is not tested IMHO.

Made sure the Tests.exe does not display the error dialog and added /all
option that does not ask interactive questions and does not wait for a
final key press.